test: alteração da estrutura dos diretórios de teste
A nova estrutura facilita manutenção e criação de novos testes.
