+++ test/usuarios/get.test.js
@@ -1,0 +1,50 @@
+const chai = require('chai')
+
+const rotaUsuarios = '/usuarios'
+const utils = require('../utils')
+
+describe(rotaUsuarios + ' GET', () => {
+  it('Query string - Busca por todos as chaves', async () => {
+    const usuario = await utils.cadastrarUsuario()
+    const { body } = await request.get(rotaUsuarios).query({
+      nome: usuario.nome,
+      email: usuario.email,
+      password: usuario.password,
+      administrador: usuario.administrador,
+      _id: usuario._id
+    }).expect(200)
+
+    chai.assert.deepEqual(body, {
+      quantidade: 1,
+      usuarios: [
+        {
+          nome: usuario.nome,
+          email: usuario.email,
+          password: usuario.password,
+          administrador: usuario.administrador,
+          _id: usuario._id
+        }
+      ]
+    })
+    await utils.excluirUsuario(usuario._id)
+  })
+
+  it('Query string - Nenhum usuário encontrado', async () => {
+    const { body } = await request.get(rotaUsuarios).query({ _id: 'a' }).expect(200)
+    chai.assert.deepEqual(body, { quantidade: 0, usuarios: [] })
+  })
+
+  it('Query string - Chave inexistente', async () => {
+    const { body } = await request.get(rotaUsuarios).query({ inexistente: 'a' }).expect(400)
+
+    chai.assert.deepEqual(body, {
+      error: {
+        name: 'ValidationError',
+        message: 'Validation Failed',
+        statusCode: 400,
+        error: 'Bad Request',
+        details: [{ inexistente: '"inexistente" is not allowed' }]
+      }
+    })
+  })
+})
